+++ FitTracker.Api/Repositories/Interfaces/IWorkoutPlanTemplateRepository.cs
@@ -1,0 +1,8 @@
+using FitTracker.Api.Models;
+
+namespace FitTracker.Api.Repositories.Interfaces;
+
+public interface IWorkoutPlanTemplateRepository
+{
+    Task<List<WorkoutPlanTemplate>> GetAllAsync();
+}